Second Night
Maybe we'll get more than 30 minutes of sleep tonight.
The project was a ton of fun. Rather than really explain the commit, the readme has all the basic information about the project now. We are submitting this commit, but we might polish some things in the morning before code freeze